[ refactor ] Terms, BankModal 리팩토링
