+++ types/bankListType.ts
@@ -1,0 +1,7 @@
+import { StaticImageData } from 'next/image';
+
+export interface BankListType {
+  name: string;
+  bankNumber: number;
+  logo: StaticImageData;
+}
